kill process

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,9 @@
-#kills a process
-exec {'killmenow':
-    path => ['usr/bin', '/usr/sbin',],
-    command => 'pkill killmenow',
-    }
+# Kill the process named killmenow
+exec { 'killmenow':
+  # Use the pkill command
+  command => 'pkill killmenow',
+  # Only run the command if the process exists
+  onlyif  => 'pgrep killmenow',
+  # Set the path for the command to execute
+  path    => '/bin:/usr/bin',
+}
